data structure update
